Add GeneDef properties handler
